css/css-values/minmax-angle-serialize.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-values-4/#comp-func">
<link rel="help" href="https://drafts.csswg.org/css-values-4/#angles">
<link rel="help" href="https://drafts.csswg.org/css-values-4/#calc-serialize">
<link rel="author" title="Xiaocheng Hu" href="mailto:xiaochengh@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../support/parsing-testcommon.js"></script>
<script>
function test_valid_angle(value, expected) {
  test_valid_value('transform', `rotate(${value})`, `rotate(${expected})`);
}

test_valid_angle('min(1deg)', 'calc(1deg)');
test_valid_angle('min(1turn)', 'calc(360deg)');
test_valid_angle('min(100grad)', 'calc(90deg)');
test_valid_angle('max(1deg)', 'calc(1deg)');
test_valid_angle('max(1turn)', 'calc(360deg)');
test_valid_angle('max(100grad)', 'calc(90deg)');
// No way to test 'rad' serialization without depending heavily on numeric serialization
// and the precision used for radians...

test_valid_angle('min(1deg, 2deg, 3deg)', 'calc(1deg)');
test_valid_angle('min(3deg, 2deg, 1deg)', 'calc(1deg)');
test_valid_angle('min(90deg, 1.58rad, 0.25turn)', 'calc(90deg)');
test_valid_angle('min(0.25turn, 1.58rad, 90deg)', 'calc(90deg)');
test_valid_angle('max(1deg, 2deg, 3deg)', 'calc(3deg)');
test_valid_angle('max(3deg, 2deg, 1deg)', 'calc(3deg)');
test_valid_angle('max(90deg, 1.57rad, 0.25turn)', 'calc(90deg)');
test_valid_angle('max(0.25turn, 1.57rad, 90deg)', 'calc(90deg)');

test_valid_angle('calc(min(1deg) + min(2deg))', 'calc(3deg)');
test_valid_angle('calc(max(1deg) + max(2deg))', 'calc(3deg)');
test_valid_angle('calc(100grad + min(1deg))', 'calc(91deg)');
test_valid_angle('calc(min(1deg) + 100grad)', 'calc(91deg)');
test_valid_angle('calc(100grad + max(1deg))', 'calc(91deg)');
test_valid_angle('calc(max(1deg) + 100grad)', 'calc(91deg)');

</script>
